Move code to SSLR
