+++ java-squid/src/main/java/org/sonar/java/ast/parser/PrinterVisitor.java
@@ -22,11 +22,11 @@ package org.sonar.java.ast.parser;
 import com.google.common.base.Charsets;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
+import com.sonar.sslr.api.typed.ActionParser;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.sonar.java.model.JavaTree;
-import org.sonar.java.parser.sslr.ActionParser;
 import org.sonar.java.resolve.JavaSymbol;
 import org.sonar.java.resolve.SemanticModel;
 import org.sonar.plugins.java.api.semantic.Type;
