+++ java-squid/src/test/java/org/sonar/java/ast/parser/PrinterVisitorTest.java
@@ -20,8 +20,8 @@
 package org.sonar.java.ast.parser;
 
 import com.google.common.base.Charsets;
+import com.sonar.sslr.api.typed.ActionParser;
 import org.junit.Test;
-import org.sonar.java.parser.sslr.ActionParser;
 import org.sonar.plugins.java.api.tree.CompilationUnitTree;
 
 import static org.fest.assertions.Assertions.assertThat;
