+++ java-squid/src/test/java/org/sonar/java/ast/parser/grammar/enums/EnumConstantTest.java
@@ -20,14 +20,14 @@
 package org.sonar.java.ast.parser.grammar.enums;
 
 import com.google.common.base.Charsets;
+import com.sonar.sslr.api.typed.ActionParser;
 import org.junit.Test;
 import org.sonar.java.ast.parser.JavaGrammar;
 import org.sonar.java.ast.parser.JavaLexer;
+import org.sonar.java.ast.parser.JavaNodeBuilder;
 import org.sonar.java.ast.parser.TreeFactory;
 import org.sonar.java.model.declaration.AnnotationTreeImpl;
 import org.sonar.java.model.declaration.EnumConstantTreeImpl;
-import org.sonar.java.parser.sslr.ActionParser;
-import org.sonar.java.parser.sslr.JavaNodeBuilder;
 import org.sonar.plugins.java.api.tree.IdentifierTree;
 import org.sonar.plugins.java.api.tree.Tree;
 import org.sonar.sslr.grammar.LexerlessGrammarBuilder;
