+++ java-squid/src/test/java/org/sonar/java/model/SyntacticEquivalenceTest.java
@@ -22,10 +22,10 @@ package org.sonar.java.model;
 import com.google.common.base.Charsets;
 import com.google.common.base.Joiner;
 import com.google.common.collect.Lists;
+import com.sonar.sslr.api.typed.ActionParser;
 import org.fest.assertions.BooleanAssert;
 import org.junit.Test;
 import org.sonar.java.ast.parser.JavaParser;
-import org.sonar.java.parser.sslr.ActionParser;
 import org.sonar.plugins.java.api.tree.ClassTree;
 import org.sonar.plugins.java.api.tree.CompilationUnitTree;
 import org.sonar.plugins.java.api.tree.MethodTree;
