+++ java-squid/src/test/java/org/sonar/java/model/expression/MethodInvocationTreeImplTest.java
@@ -21,10 +21,10 @@ package org.sonar.java.model.expression;
 
 import com.google.common.base.Charsets;
 import com.google.common.collect.Lists;
+import com.sonar.sslr.api.typed.ActionParser;
 import org.fest.assertions.Assertions;
 import org.junit.Test;
 import org.sonar.java.ast.parser.JavaParser;
-import org.sonar.java.parser.sslr.ActionParser;
 import org.sonar.java.resolve.SemanticModel;
 import org.sonar.plugins.java.api.semantic.Symbol;
 import org.sonar.plugins.java.api.tree.ClassTree;
