+++ java-squid/src/test/java/org/sonar/sslr/tests/Assertions.java
@@ -23,15 +23,15 @@ import com.google.common.base.Charsets;
 import com.google.common.base.Preconditions;
 import com.sonar.sslr.api.RecognitionException;
 import com.sonar.sslr.api.Rule;
+import com.sonar.sslr.api.typed.ActionParser;
 import org.fest.assertions.GenericAssert;
 import org.sonar.java.ast.parser.FormalParametersListTreeImpl;
 import org.sonar.java.ast.parser.JavaGrammar;
 import org.sonar.java.ast.parser.JavaLexer;
+import org.sonar.java.ast.parser.JavaNodeBuilder;
 import org.sonar.java.ast.parser.TreeFactory;
 import org.sonar.java.model.InternalSyntaxToken;
 import org.sonar.java.model.JavaTree;
-import org.sonar.java.parser.sslr.ActionParser;
-import org.sonar.java.parser.sslr.JavaNodeBuilder;
 import org.sonar.java.syntaxtoken.LastSyntaxTokenFinder;
 import org.sonar.plugins.java.api.tree.Tree;
 import org.sonar.sslr.grammar.GrammarRuleKey;
