+++ pom.xml
@@ -71,7 +71,7 @@
     <maven.test.redirectTestOutputToFile>true</maven.test.redirectTestOutputToFile>
 
     <sonar.version>4.5.2</sonar.version>
-    <sslr.version>1.20</sslr.version>
+    <sslr.version>1.21-SNAPSHOT</sslr.version>
     <jacoco.previous.version>0.7.4.201502262128</jacoco.previous.version>
     <jacoco.version>0.7.5.201505241946</jacoco.version>
     <asm.version>5.0.3</asm.version>
